Småändringar. Delat upp koden ytterliggare filer för lättare överblick. Klar för inlämning.
